nydus: fix a few clippy errors
Replace manual modulo operations with the more readable is_multiple_of() method
for alignment validation throughout the codebase. This improves code clarity and
maintains the same functionality.

Update error handling to use io::Error::other() instead of manually constructing
errors with ErrorKind::Other, following Rust best practices.

Fix clippy warnings by boxing large enum variants and removing unnecessary mut
qualifiers.

+++ storage/src/backend/pauser.rs
@@ -305,7 +305,7 @@ mod tests {
             let pauser = pauser.clone();
             let handle = thread::spawn(move || {
                 thread::sleep(Duration::from_millis(i * 10));
-                pauser.set_pause(Duration::from_millis(100 + i as u64 * 10));
+                pauser.set_pause(Duration::from_millis(100 + i * 10));
             });
             handles.push(handle);
         }
